drm/omap: Hide "cast from pointer to integer of different size"
Get rid of "warning: cast from pointer to integer of different
size". The pointer in 64-bit system is 64-bit. Casting first to
integer of matching integer size (ptrdiff_t) and only then to final
uint32_t makes the compiler happy. The "pointer" is really not a
pointer at all so there is no risk of overflow.

+++ drivers/gpu/drm/omapdrm/omap_drv.c
@@ -545,7 +545,7 @@ static int pdev_probe(struct platform_device *pdev)
 	priv->dispc_ops = dispc_get_ops();
 
 	soc = soc_device_match(omapdrm_soc_devices);
-	priv->omaprev = soc ? (unsigned int)soc->data : 0;
+	priv->omaprev = soc ? (uint32_t)(ptrdiff_t)soc->data : 0;
 	priv->wq = alloc_ordered_workqueue("omapdrm", 0);
 
 	spin_lock_init(&priv->list_lock);
